MET-504: move enrichment/dereference code to more appropriate projects.
Firstly: there is no need to have client code (utils)  in the various
common projects: the services don't need this code at all. Secondly: we
can merge the very small dereference client project into the enrichment
client project.
