--- metis-dereference/metis-dereference-client/.gitignore
@@ -1,3 +1,0 @@
-##Add to ignore to not commit by mistake
-/src/main/resources/client.properties
-/src/test/resources/client.properties
