--- metis-dereference/metis-dereference-client/pom.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-		 Copyright 2007-2013 The Europeana Foundation
-
-		  Licenced under the EUPL, Version 1.1 (the "Licence") and subsequent versions as approved
-		  by the European Commission;
-		  You may not use this work except in compliance with the Licence.
-
-		  You may obtain a copy of the Licence at:
-		  http://joinup.ec.europa.eu/software/page/eupl
-
-		  Unless required by applicable law or agreed to in writing, software distributed under
-		  the Licence is distributed on an "AS IS" basis, without warranties or conditions of
-		  any kind, either express or implied.
-		  See the Licence for the specific language governing permissions and limitations under
-		  the Licence.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>metis-dereference</artifactId>
-        <groupId>eu.europeana.metis</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>metis-dereference-client</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>eu.europeana.metis</groupId>
-            <artifactId>metis-dereference-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-            <version>${version.spring}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${version.jackson}</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-io</groupId>
-            <artifactId>commons-io</artifactId>
-            <version>${version.commons.io}</version>
-        </dependency>
-        <dependency>
-            <groupId>eu.europeana.metis</groupId>
-            <artifactId>metis-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>eu.europeana.metis</groupId>
-            <artifactId>metis-enrichment-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
