--- metis-dereference/metis-dereference-client/src/test/resources/client.properties.example
@@ -1,1 +1,0 @@
-host.url=
