+++ metis-dereference/metis-dereference-common/pom.xml
@@ -32,27 +32,6 @@
             <artifactId>morphia</artifactId>
             <version>${version.morphia}</version>
         </dependency>
-        <dependency>
-            <groupId>eu.europeana.metis</groupId>
-            <artifactId>metis-enrichment-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-		<dependency>
-			<groupId>eu.europeana.corelib</groupId>
-			<artifactId>corelib-edm-definitions</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.mockito</groupId>
-			<artifactId>mockito-core</artifactId>
-			<version>${version.mockito.core}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>${version.junit}</version>
-			<scope>test</scope>
-		</dependency>
     </dependencies>
 
 </project>
