+++ metis-dereference/metis-dereference-common/src/main/java/eu/europeana/metis/dereference/OriginalEntity.java
@@ -16,7 +16,6 @@
  */
 package eu.europeana.metis.dereference;
 
-
 import org.mongodb.morphia.annotations.Entity;
 import org.mongodb.morphia.annotations.Id;
 import org.mongodb.morphia.annotations.Indexed;
