+++ metis-dereference/metis-dereference-common/src/main/java/eu/europeana/metis/dereference/Vocabulary.java
@@ -16,11 +16,8 @@
  */
 package eu.europeana.metis.dereference;
 
-
 import java.io.Serializable;
-
 import javax.xml.bind.annotation.XmlElement;
-
 import org.mongodb.morphia.annotations.Entity;
 import org.mongodb.morphia.annotations.Id;
 import org.mongodb.morphia.annotations.Indexed;
