+++ metis-dereference/pom.xml
@@ -32,7 +32,6 @@
         <module>metis-dereference-common</module>
         <module>metis-dereference-service</module>
         <module>metis-dereference-rest</module>
-        <module>metis-dereference-client</module>
     </modules>
     <build>
         <plugins>
