+++ metis-enrichment/metis-enrichment-client/pom.xml
@@ -43,6 +43,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>eu.europeana.metis</groupId>
+            <artifactId>metis-dereference-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 		<dependency>
 		    <groupId>org.slf4j</groupId>
 		    <artifactId>slf4j-simple</artifactId>
@@ -146,12 +151,6 @@
             <groupId>eu.europeana.metis</groupId>
             <artifactId>metis-common</artifactId>
             <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-        	<groupId>eu.europeana.metis</groupId>
-        	<artifactId>metis-dereference-client</artifactId>
-        	<version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
     <build>
